requirements.txt
eth-brownie>=1.11.0,<2.0.0
python-dotenv>=0.21.0